move test file next to source
