+++ routes/experiences/index.test.js
@@ -3,21 +3,21 @@ chai.use(require('chai-datetime'));
 
 const assert = chai.assert;
 const request = require('supertest');
-const app = require('../../../app');
+const app = require('../../app');
 const {
     MongoClient,
     ObjectId,
 } = require('mongodb');
 const config = require('config');
 const sinon = require('sinon');
-const authentication = require('../../../libs/authentication');
+const authentication = require('../../libs/authentication');
 const {
     generateInterviewExperienceData,
     generateWorkExperienceData,
-} = require('../testData');
+} = require('./testData');
 
-const create_company_keyword_collection = require('../../../database/migrations/create-companyKeywords-collection');
-const create_title_keyword_collection = require('../../../database/migrations/create-jobTitleKeywords-collection');
+const create_company_keyword_collection = require('../../database/migrations/create-companyKeywords-collection');
+const create_title_keyword_collection = require('../../database/migrations/create-jobTitleKeywords-collection');
 
 describe('Experiences 面試和工作經驗資訊', () => {
     let db;
