+++ routes/index.test.js
@@ -8,6 +8,7 @@ const request = require('supertest');
 const app = require('../app');
 const authentication = require('../libs/authentication');
 const authorization = require('../libs/authorization');
+const recommendation = require('../libs/recommendation');
 
 describe('GET /me/permission/search 確認使用者查詢資訊權限', () => {
     let sandbox;
@@ -67,6 +68,59 @@ describe('GET /me/permission/search 確認使用者查詢資訊權限', () => {
 
                 assert.propertyVal(res.body, 'hasSearchPermission', false);
             });
+    });
+
+    afterEach(() => {
+        sandbox.restore();
+    });
+});
+
+describe('POST /me/recommendations 取得使用者推薦字串', () => {
+    let sandbox;
+
+    beforeEach(() => {
+        sandbox = sinon.sandbox.create();
+    });
+
+    it('get recommendation string success', () => {
+        const cachedFacebookAuthentication = sandbox.stub(authentication, 'cachedFacebookAuthentication')
+            .withArgs(sinon.match.object, sinon.match.object, 'fakeaccesstoken')
+            .resolves({ facebook_id: '-1' });
+
+        const getRecommendationString = sandbox.stub(recommendation, 'getRecommendationString')
+            .withArgs(sinon.match.object, { id: '-1', type: 'facebook' })
+            .resolves('00000000ccd8958909a983e9');
+
+        return request(app).post('/me/recommendations')
+            .send({
+                access_token: 'fakeaccesstoken',
+            })
+            .expect(200)
+            .expect((res) => {
+                sinon.assert.calledOnce(cachedFacebookAuthentication);
+                sinon.assert.calledOnce(getRecommendationString);
+
+                assert.propertyVal(res.body, 'recommendation_string', '00000000ccd8958909a983e9');
+            });
+    });
+
+    it('fail if facebook auth fail', () => {
+        sandbox.stub(authentication, 'cachedFacebookAuthentication').rejects();
+
+        return request(app).post('/me/recommendations')
+            .expect(401);
+    });
+
+    it('fail if getRecommendationString fail', () => {
+        sandbox.stub(authentication, 'cachedFacebookAuthentication')
+            .resolves({ _id: { id: '-1', type: 'facebook' } });
+        sandbox.stub(recommendation, 'getRecommendationString').rejects();
+
+        return request(app).post('/me/recommendations')
+            .send({
+                access_token: 'fakeaccesstoken',
+            })
+            .expect(500);
     });
 
     afterEach(() => {
